cleaning up beacon chain helper methods
