+++ test/_lib/utils/general.js
@@ -1,5 +1,4 @@
 const moment = require('moment');
-const ssz = require('@chainsafesystems/ssz');
 import { RocketStorage } from '../artifacts'
 
 // The newer version of Web3. Waiting for them to upgrade truffles web3.
@@ -186,44 +185,5 @@ export function getTransactionContractEvents(txResult, contractAddress, eventNam
             if (decodeParam.indexed && (decodeParam.type == 'string' || decodeParam.type == 'bytes')) decodeParam.type = 'bytes32'; // Issues decoding indexed string and bytes parameters
             return decodeParam;
         }), log.data, log.topics.slice(1)));
-}
-
-
-// Create a serialised DepositInput object from a given set of credentials
-export function getDepositInput({pubkey, withdrawalPubkey}) {
-
-    // Default credentials
-    if (!pubkey) pubkey = '0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef';
-    if (!withdrawalPubkey) withdrawalPubkey = '0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef';
-
-    // Get withdrawal credentials
-    let withdrawalCredentials = Buffer.concat([
-        Buffer.from('00', 'hex'), // BLS_WITHDRAWAL_PREFIX_BYTE
-        Buffer.from(web3.utils.sha3(withdrawalPubkey).substr(2), 'hex').slice(1) // Last 31 bytes of withdrawal pubkey hash
-    ], 32);
-
-    // Get proof of possession
-    // TODO: implement correctly once BLS library found
-    let proofOfPossession = Buffer.from(
-        '0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef' +
-        '0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef' +
-        '0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef',
-        'hex'
-    );
-
-    // Return depositInput
-    return ssz.serialize(
-        {
-            'pubkey': Buffer.from(pubkey, 'hex'),
-            'withdrawal_credentials': withdrawalCredentials,
-            'proof_of_possession': proofOfPossession,
-        },
-        {fields: [
-            ['pubkey', 'bytes48'],
-            ['withdrawal_credentials', 'bytes32'],
-            ['proof_of_possession', 'bytes96'],
-        ]}
-    );
-
 }
